Set up SPoRC model weighting

Set up SPoRC model weighting

Usage

Setup_Mod_Weighting(
  input_list,
  sablefish_ADMB,
  likelihoods,
  Wt_Catch = 1,
  Wt_FishIdx = 1,
  Wt_SrvIdx = 1,
  Wt_Rec = 1,
  Wt_F = 1,
  Wt_Tagging = 1,
  Wt_FishAgeComps,
  Wt_SrvAgeComps,
  Wt_FishLenComps,
  Wt_SrvLenComps
)

Arguments

input_list

List containing data, parameter, and map lists.

sablefish_ADMB

Numeric flag to mimic calculations for the sablefish ADMB model:

1 to mimic sablefish ADMB calculations

0 otherwise

likelihoods

Numeric flag indicating likelihood implementation to use:

0 for ADMB likelihoods

1 for TMB likelihoods

Wt_Catch

Either a numeric scalar (lambda) applied to the overall catch dataset or an array of lambdas (i.e., weights can change by year and fleet) dimensioned by n_regions, n_years, n_fish_fleets.

Wt_FishIdx

Either a numeric scalar (lambda) applied to the overall fishery index dataset or an array of lambdas (i.e., weights can change by year and fleet) dimensioned by n_regions, n_years, n_fish_fleets.

Wt_SrvIdx

Either a numeric scalar (lambda) applied to the overall survey index dataset or an array of lambdas (i.e., weights can change by year and fleet) dimensioned by n_regions, n_years, n_srv_fleets.

Wt_Rec

Numeric weight (lambda) applied to the recruitment penalty.

Wt_F

Numeric weight (lambda) applied to the fishing mortality penalty.

Wt_Tagging

Numeric weight (lambda) applied to tagging data.

Wt_FishAgeComps

Numeric weight (lambda) applied to fishery age composition data.

Wt_SrvAgeComps

Numeric weight (lambda) applied to survey age composition data.

Wt_FishLenComps

Numeric weight (lambda) applied to fishery length composition data.

Wt_SrvLenComps

Numeric weight (lambda) applied to survey length composition data.