Get Retrospective Plot

Get Retrospective Plot

Usage

get_retrospective_plot(retro_output, Rec_Age)

Arguments

retro_output

Dataframe generated from do_retrospective

Rec_Age

Age in which recruitment occurs

Value

A retrospective plot of recruitment and SSB in relative and absolute scales, as well as a retrospective plot of recruitment by cohort (squid plot)

See also

Examples

if (FALSE) { # \dontrun{
# do retrospective
retro <- do_retrospective(n_retro = 7, # number of retro peels to run
data = data, # rtmb data
parameters = parameters, # rtmb parameters
mapping = mapping, # rtmb mapping
random = NULL, # if random effects are used
do_par = TRUE, # whether or not to parralleize
n_cores = 7, # if parallel, number of cores to use
do_francis = F, # if we want tod o Francis
n_francis_iter = NULL # Number of francis iterations to do
)
get_retrospective_plot(retro, Rec_Age = 2)
} # }