Plots OSA residuals from outputs from get_osa. Much of this code is taken from the afscOM package, but with modificaitons to plot features.

Plots OSA residuals from outputs from get_osa. Much of this code is taken from the afscOM package, but with modificaitons to plot features.

Usage

plot_resids(osa_results)

Arguments

osa_results

List object obtained from get_osa, that contains a dataframe of residuals and aggregated fits.

Value

A vareity of plots for OSA residuals (list)

See also

Examples

if (FALSE) { # \dontrun{
comp_props <- get_comp_prop(data = data, rep = sabie_rtmb_model$rep,
 age_labels = 2:31, len_labels = seq(41, 99, 2), year_labels = 1960:2024)
plot_resids(get_osa(obs_mat = comp_props$Obs_FishAge_mat,
                    exp_mat = comp_props$Pred_FishAge_mat,
                    N = rep(16.52215, length(1999:2023)),
                    years = which(1960:2024 %in% 1999:2023),
                    LN_Sigma = LN_Sigma,
                    fleet = 1,
                    bins = 2:31,
                    comp_type = 0,
                    comp_like = 0,
                    bin_label = "Age"))
osa_plot <- plot_resids(osa_results)
} # }