Setup biological inputs for estimation model

Setup biological inputs for estimation model

Usage

Setup_Mod_Biologicals(
  input_list,
  WAA,
  MatAA,
  AgeingError = NULL,
  Use_M_prior = 0,
  M_prior = NA,
  fit_lengths = 0,
  SizeAgeTrans = NA,
  Selex_Type = "age",
  M_spec = NULL,
  Fixed_natmort = NULL,
  ...
)

Arguments

input_list

List containing data, parameter, and map lists for the model.

WAA

Numeric array of weight-at-age, dimensioned [n_regions, n_years, n_ages, n_sexes].

MatAA

Numeric array of maturity-at-age, dimensioned [n_regions, n_years, n_ages, n_sexes].

AgeingError

Numeric matrix representing the ageing error transition matrix, dimensioned by [number of modeled ages, number of observed composition ages]. Defaults to identity matrix if not specified (no ageing error).

Use_M_prior

Integer flag indicating whether to apply a natural mortality prior (0 = no, 1 = yes).

M_prior

Numeric vector of length two giving the mean (in normal space) and standard deviation of the natural mortality prior.

fit_lengths

Integer flag indicating whether to fit length data (0 = no, 1 = yes).

SizeAgeTrans

Numeric array of size-at-age transition probabilities, dimensioned [n_regions, n_years, n_lens, n_ages, n_sexes].

Selex_Type

Character string specifying whether selectivity is age or length-based. Default is age-based

"length": Length-based selectivity.

"age": Age-based selectivity

M_spec

Character string specifying natural mortality estimation approach. Defaults to NULL, which estimates mortality for each sex independently. Other options:

"est_shared_s": Estimate a single natural mortality rate shared across sexes (if n_sexes == 2).

"fix": Fix all natural mortality parameters using the provided array.

Fixed_natmort

Numeric array of fixed natural mortality values, dimensioned [n_regions, n_years, n_ages, n_sexes]. Required if M_spec = "fix".

...

Additional arguments for starting values such as ln_M and M_offset. These are ignored if M_spec = "fix".