Initialize Simulation Dimension Settings

Creates and returns a list of key dimension values used to set up a simulation or management strategy evaluation (MSE). This list provides structural information such as number of simulations, years, regions, ages, fleets, and whether to include a feedback loop.

Usage

Setup_Sim_Dim(
  n_sims,
  n_yrs,
  n_regions,
  n_ages,
  n_lens,
  n_obs_ages = n_ages,
  n_sexes,
  n_fish_fleets,
  n_srv_fleets,
  run_feedback = FALSE,
  feedback_start_yr = NULL
)

Arguments

n_sims

Integer. Number of simulation replicates.

n_yrs

Integer. Number of years in the simulation.

n_regions

Integer. Number of modeled regions.

n_ages

Integer. Number of modeled age classes.

n_lens

Integer. Number of modeled length bins.

n_obs_ages

Integer. Number of observed age classes (can differ from n_ages, default = n_ages).

n_sexes

Integer. Number of sexes.

n_fish_fleets

Integer. Number of fishery fleets.

n_srv_fleets

Integer. Number of survey fleets.

run_feedback

Logical. Whether to include a feedback management loop (default = FALSE).

feedback_start_yr

Integer. First year that feedback is applied (only used if run_feedback = TRUE).

Value

A list containing the specified dimension values, with elements:

n_sims, n_yrs, n_regions, n_ages, n_lens, n_obs_ages, n_sexes, n_fish_fleets, n_srv_fleets

init_iter (set internally to n_ages * 10)

feedback_start_yr, run_feedback

See also